requirements.txt
astroid==1.6.0
defusedxml==0.5.0
Django==1.11.6
django-ckeditor==5.3.1
django-debug-toolbar==1.8
django-filter==1.1.0
django-js-asset==0.1.1
django-ratelimit==1.1.0
django-sslify-admin==0.5.0
django-sslserver==0.20
django-suit==0.2.25
django-tinymce==2.6.0
djangocms-admin-style==1.2.7
djangorestframework==3.7.1
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
pylint==1.8.1
pytz==2017.2
PyYAML==3.12
six==1.11.0
sqlparse==0.2.4
whitenoise==3.3.1
wrapt==1.10.11
